react refactor
